coal: Fix permalink loading for default period
Follows-up 1101cd8.

When loading the page, the hash is updated to reflect the default
period. Reloading this url resulted in a blank page because init()
and selectTab() had incompatible assumptions about the class being
set in the HTML.

Fix by only calling selectTab() for one of the other periods (e.g.
day, week etc.). If the permalink is for hour, call drawCharts()
directly since the class is already set.

As a bonus, when accessing a non-existent period, it now falls
back to the default instead of loading nothing.

(cherry picked from commit a1233047a742e2140d74d7119133180e45f9e156)

Change-Id: I6e7f4fdc6ba0ec623f7ac7ab7ef28433d7a76722