Show yearly graphs
Support for a one-year range was built into coal / coal-web from the beginning
(try https://performance.wikimedia.org/coal/v1/metrics?period=year) but I did
not include it in the UI. But now that we already have eight months' worth of
data, it makes sense to turn it on.

The section of the graph that represents the time before eight months' ago will
have the same value as the first real measurement. So the graph will simply be
artificially flat in that part. I think it is not too misleading.

Change-Id: I488a2178d27554a481d3aff6ef45bb653d94f846