Make coal-web server address configurable
Change the default from <https://performance.wikimedia.org/coal> to `/coal`.
The default has to be something that works in both production and Beta,
where it is always on the same public domain, mounted via mod_proxy at /coal.

The downside of this new default is that when working on the site locally,
the Jekyll server won't serve `/coal`, resulting in a 404 Not Found.

Address this by making it configurable. You can choose to either
set it to <https://performance.wikimedia.org/coal> to use the live instance,
or something like <http://127.0.0.1:5000> to use a local instance.

The 'npm run generate' build for pre-commit, always uses `/coal`.

The 'npm run jekyll' build that creates a local server, leaves it unset,
allowing it to be set by setting COAL_WEB_SERVER, with a default of
<https://performance.wikimedia.org/coal> for convenience in env.rb.

Bug: T195314
Change-Id: Id3e98d0a556fa184c76bae951349e08b65496dc3