build: Update development dependencies
Before:
> 8 vulnerabilities found [645 packages audited]
>   Severity: 1 Low | 6 Moderate | 1 High

After:
> [+] no known vulnerabilities found [491 packages audited]

Difference:
> eslint                   ^3.18.0  →  ^4.19.1
> eslint-config-wikimedia    0.4.0  →    0.5.0
> uncss                     0.14.1  →   0.16.2

The 'uncss' package was written in 0.15.0 to use jsdom instead of PhantomJS.
Among other things, this means it no longer runs a semi-real browser in a
seperate process, but rather emulates a DOM within Node.js. The way it is
implemented, synchronously loads scripts, styles and image references when
it builds its tree, and hard-fails if any of the referenced files are unknown.
This is a problem, because when jsdom opens the HTML file statically over
the file:// protocol, the relative "/coal" reference is interpreted as file
"/coal"  in the root directory of the disk, which doesn't exist. Work around
this issue by doing a temporary jekyll build right before the uncss task that
doesn't have the hardcoded "/coal" url set.

The eslint update detected a new indentation errors, which were fixed.

Change-Id: I17fc5b16e4da896944ff6a24199f2f73a8793bfd