build: Specify version of Jekyll and encourage non-global install
Use a Gemfile to specify the version of Jekyll so that we're all
using the same version. Also document use of bundler so that it
doesn't require a global install (which makes it difficult to
work with other projects), and also to make it easy to add plugins
(see next commit).

Change-Id: Id2978f7751337b4c35e28816df998c44ca799090