build: Enable use of _config.yml file
The file wasn't used previously. It seems that if --source or
--destination is passed by CLI, then the config file isn't used
(instead of e.g. merging/overriding). There's probably a way to
make overrides work, but we don't need that, it was duplicating
the value.

The way I found out is that in the next commit I add a 'permalink'
key `_config.yml`, which didn't work unless I removed the CLI
params.

Also, re-order the script keys in package.json while at it to
be more "define before use"-ish.

Change-Id: Iabc534ca9df04d5586a2a74bb1b8ea913c5bf380