build: Upgrade to Jekyll 4
Update from 3.x to 4.x, this is in preparation for the switch
from Bootstrap CSS to Wikimedia UI so I don't have to migrate
the plugins that involves.

Everything is compatible between 3.x and 4.x for us. The only
observable difference seems to be the order of some of the
HTML tags that the static redirect pages involves, which seems
harmless.

Bug: T260752
Change-Id: I6392a9c9b2fd2bc4371b27121a508c1351bccaba