Remove `COAL_WEB_SERVER` complexity
This only existed so that <https://performance.wikimedia.beta.wmflabs.org/>
would use its local Coal instance instead of the prod one, however
I don't think that's worth the complexity. If between now and removal
of Coal we need to test that in Beta, we can test it directly by
opening the /coal URL in the browser or with Curl.

This is in prep for moving from the Ruby/Jekyll stack to something
simpler and more standalone based on PHP, allowing for more re-use
between doc.wikimedia.org, security.wikimedia.org,
research.wikimedia.org, and performance.wikimedia.org.

Change-Id: I5a2d30c68974ee666ae2ca166a0c7cffd3d029ff