php-profiling: Remove unused apaxy header/footer
These have not worked for several years, because the flamegraphs are
served from a separate server than the perf.wm.o site, and so Apache
doesn't see them. This is fine, we don't actually need this level of
customisation. It seems valuable for our Apache file listing to be
more standard.

The broken config is removed from puppet.git with change I8d392a6b74.

Change-Id: I00820f66385d69d77b1c09a98d2f1a35f4ee2f05