php-profiling: Create new flame graph selectors
== What ==

* Date: You can now choose any date in the last 2+ years, not 5 days.
  By choosing "Other" you get a native date selector for any date
  between 2021-03-1 and today.

* Source: Kubernetes performance data is now available through here
  (excimer-k8s and excimer-k8s-wall). The default value remains
  baremetal because that is where >99% of traffic is today.

* Entrypoint: I'v added JobQueue runners as well as our focussed
  precision datasets for EditAction, PreSend, and PostSend that
  we added in T253679. These were previously available via the
  directory browser already, but now more easily discoverable through
  here.

== How ==

The previous listing was generated in JavaScript and optional. The
links to the directory work either way, the JS code is only for
finding the links to SVG files more quickly. It generated a few
SVG URLs with the default settings and dumped them into a bullet
list.

The new interface is an HTML form with a more streamlined flow
from top to bottom, guiding you through the various choices. This way
we can explose the full range of flamegraphs we generate (and store
in the underlying file directory) without an overwhelming interface
since the interface has a fixed number of choices. We don't have to
expose all possible combinations at once, which previously motivated
us to hide most of the variants, thus making them only discoverable
via the file browser.

As before, we still start with "good" defaults, so you don't have
to make any choices to start exploring. You can press the big blue
button right away to get the most recent index.php flamegraph with the
default settings.

The form is hidden by default (same as before, the file browser links
are available without JS), and made visible once the JS code has
filled in the dynamic date-based values.

Whenever any input field changes, the JS code responds by making the
image, link, and button point to the SVG flamegraph for the now-selected
date, source, and other settings.

Note that while the file browser at https://performance.wikimedia.org/arclamp/
lists the files that exist directly on the arclamp1001 server, which
is limited to the last few weeks only. The form offers you the range
back to 2021. This is possible because the Apache server that serves
the SVG files actually proxies the `*.svgz` URLs to Swift. [1] Thus
after we delete the gigabytes of raw trace logs and flame graphs from
the serve locallly, the flamegraph URLs keep working.

[1] https://gerrit.wikimedia.org/g/operations/puppet/+/0c84a5d3b6/modules/profile/templates/webperf/site/performance-website.erb#46

Bug: T200108
Change-Id: I4986914086fa77da641054b233de6b86f9ac7310