php-profiling: Fix race condition in <select> on page refresh
When refreshing the page, the browser's native memory for <form>
value conflicted in the date field. It would try to find e.g.
2023-04-13 and find it wasn't there at first, and so it would fallback
to selecting the first option, "today", instead of our default of
"yesterday". It works fine on hard refresh, but on regular refresh
it would break in this way.

It happened because the JS code preserved the <option> element
and its selected status, but it changed the value="" key. My mistake
was assuming the browser remembers the offset rather than the key.

The benefit of this is that if one changes the order or length of
menu at any time, browser's form memory applies properly after major
changes in some cases.

Change-Id: If73bcf2376433e82dd357a883e39ca857d548101