asreport: Fix unstable interpretation of asreport.tsv date
This bug has been around for many years. Whenever the 'renew-asreport'
command is run, we download the 'latest' TSV file, which (correctly)
also sets the file mtime to match the Last-Modified date of the HTTP
resource. Thus the person updating the report, will correctly generate
the page with the date (not today's date, but the date of when the
report data was published, typically the first day of a month).

However, when someone else checks out the code from Git, their files
have the mtime of "today" because that is how Git works. File modification
times are not part of the file metadata that Git manages. The person
making the change originally may also be affected if they e.g. switch
between branches etc.

Fix this by keeping the date in a separate file.

This commit stores it as 2023-01-01. After trying the last few
files at https://analytics.wikimedia.org/published/datasets/performance/autonomoussystems/
one by one, I found that the current published report is in fact that
of January. The date updates are unrelated and caused by this bug.

Change-Id: Ibe556738715b7827b1ceb31d6e4479e4ca3a5221